refactor: decompose 3 more monolith files into submodules (issue #2486 sprint 2)
Continues sprint-based decomposition of files exceeding 500 LOC, following
the backward-compatible coordinator pattern from sprint 1.

- signal_processing.py (982 → 59 LOC) → _spectral_analysis, _time_domain,
  _wavelet, _dtw, _filters submodules
- flexible_shaft.py (1056 → 77 LOC) → _shaft_data, _shaft_fem, _shaft_model,
  _shaft_properties, _shaft_utils submodules
- inverse_dynamics.py (979 → 38 LOC) → _id_core, _id_models, _id_export submodules

All public symbols remain importable from the original module paths.
