+++ src/engines/physics_engines/mujoco/python/mujoco_humanoid_golf/inverse_dynamics.py
@@ -12,968 +12,31 @@ torques from desired motion. Includes:
 
 from __future__ import annotations
 
-import csv
-from dataclasses import dataclass
-
-import mujoco
-import numpy as np
-from scipy.linalg import lstsq
-
-from src.shared.python.core.contracts import precondition
 from src.shared.python.logging_pkg.logging_config import get_logger
 
-from .kinematic_forces import KinematicForceAnalyzer, MjDataContext
+from ._id_core import InverseDynamicsSolver, RecursiveNewtonEuler
+from ._id_export import (
+    InverseDynamicsAnalyzer,
+    _build_inverse_dynamics_csv_row,
+    _validate_inverse_dynamics_export_inputs,
+    export_inverse_dynamics_to_csv,
+)
+from ._id_models import (
+    ForceDecomposition,
+    InducedAccelerationResult,
+    InverseDynamicsResult,
+)
 
 logger = get_logger(__name__)
 
-
-@dataclass
-class InducedAccelerationResult:
-    """Result of induced acceleration analysis."""
-
-    gravity: np.ndarray  # Acceleration induced by gravity
-    velocity: np.ndarray  # Acceleration induced by velocity (Coriolis/Centrifugal)
-    control: np.ndarray  # Acceleration induced by control torques
-    total: np.ndarray  # Total acceleration
-
-
-@dataclass
-class InverseDynamicsResult:
-    """Result of inverse dynamics computation."""
-
-    joint_torques: np.ndarray  # [nv] - Required joint torques
-    constraint_forces: np.ndarray | None = (
-        None  # Constraint forces (if parallel mechanism)
-    )
-
-    # Force decomposition
-    inertial_torques: np.ndarray | None = None  # Ma term
-    coriolis_torques: np.ndarray | None = None  # C(q,q̇)q̇ term
-    gravity_torques: np.ndarray | None = None  # g(q) term
-
-    # Task-space forces
-    end_effector_force: np.ndarray | None = None  # Force at end-effector
-
-    # Validation metrics
-    residual_norm: float = 0.0  # For least-squares solutions
-    is_feasible: bool = True  # Whether solution is physically feasible
-
-    # Phase 4 / Advanced Control additions
-    success: bool = True
-    manipulability_index: float | None = None
-    joint_names: list[str] | None = None
-
-
-@dataclass
-class ForceDecomposition:
-    """Decomposition of forces/torques into components."""
-
-    total: np.ndarray  # Total force/torque
-    inertial: np.ndarray  # Due to acceleration (Ma)
-    coriolis: np.ndarray  # Due to velocity (C(q,q̇)q̇)
-    centrifugal: np.ndarray  # Due to centrifugal effects
-    gravity: np.ndarray  # Due to gravity (g(q))
-    external: np.ndarray | None = None  # External forces
-
-
-class InverseDynamicsSolver:
-    """Solve inverse dynamics for golf swing models.
-
-    This class computes the joint torques required to achieve a desired
-    motion trajectory. Handles both open-chain and closed-chain (parallel
-    mechanism) systems.
-
-    Key Methods:
-    - solve_inverse_dynamics(): Main method for full trajectory
-    - compute_required_torques(): Single time step
-    - decompose_forces(): Break down torques into components
-    """
-
-    def __init__(self, model: mujoco.MjModel, data: mujoco.MjData) -> None:
-        """Initialize inverse dynamics solver.
-
-        Args:
-            model: MuJoCo model
-            data: MuJoCo data
-        """
-        if not (model is not None):
-            raise ValueError("model must be provided")
-        if not (model is not None):
-            raise ValueError("model must be provided")
-        self.model = model
-        self.data = data
-
-        # Initialize kinematic force analyzer
-        self.kinematic_analyzer = KinematicForceAnalyzer(model, data)
-
-        # Check if model has constraints (parallel mechanism)
-        self.has_constraints = (model.neq > 0) or self._detect_closed_chains()
-
-        # Optimization: Pre-allocate Jacobian arrays and detect API signature
-        # This avoids try-except overhead in tight loops
-        # (e.g. compute_end_effector_forces)
-        self._use_flat_jacobian = False
-        try:
-            # Test with dummy arrays to check signature
-            # Body 0 (world) is always valid
-            jacp = np.zeros((3, model.nv))
-            jacr = np.zeros((3, model.nv))
-            mujoco.mj_jacBody(model, data, jacp, jacr, 0)
-            self._jacp = jacp
-            self._jacr = jacr
-        except (TypeError, ValueError):
-            # Fallback to flat array approach for older MuJoCo bindings
-            self._use_flat_jacobian = True
-            self._jacp_flat = np.zeros(3 * model.nv)
-            self._jacr_flat = np.zeros(3 * model.nv)
-
-        # CRITICAL FIX (Phase 1): Dedicated MjData for thread-safe physics
-        # Prevents "Observer Effect" where analysis corrupts visualization state
-        self._perturb_data = mujoco.MjData(model)
-
-    def _detect_closed_chains(self) -> bool:
-        """Detect if model has closed kinematic chains.
-
-        Returns:
-            True if closed chains detected
-        """
-        # Simple heuristic: Check for equality constraints
-        # In production, more sophisticated analysis needed
-        return bool(self.model.neq > 0)
-
-    def compute_required_torques(
-        self,
-        qpos: np.ndarray,
-        qvel: np.ndarray,
-        qacc: np.ndarray,
-        external_forces: np.ndarray | None = None,
-    ) -> InverseDynamicsResult:
-        """Compute required joint torques for desired motion.
-
-        Uses the equation of motion:
-        M(q)q̈ + C(q,q̇)q̇ + g(q) = τ + τ_ext
-
-        Args:
-            qpos: Joint positions [nv]
-            qvel: Joint velocities [nv]
-            qacc: Joint accelerations [nv]
-            external_forces: External forces [nv] (optional)
-
-        Returns:
-            InverseDynamicsResult with computed torques
-        """
-        # Set state (Thread-Safe: use private data)
-        if not (qpos is not None):
-            raise ValueError("qpos must be provided")
-        if not (qpos is not None):
-            raise ValueError("qpos must be provided")
-        self._perturb_data.qpos[:] = qpos
-        self._perturb_data.qvel[:] = qvel
-        self._perturb_data.qacc[:] = qacc
-
-        # Forward kinematics and dynamics
-        # This computes qfrc_bias = C(q,q̇)q̇ + g(q)
-        mujoco.mj_forward(self.model, self._perturb_data)
-
-        # Compute inverse dynamics
-        # This computes qfrc_inverse = M(q)q̈ + C(q,q̇)q̇ + g(q) - ext
-        mujoco.mj_inverse(self.model, self._perturb_data)
-
-        qfrc_inverse = self._perturb_data.qfrc_inverse.copy()
-
-        # Decompose if needed (optional, for result detail)
-        # For now, just return total
-
-        return InverseDynamicsResult(
-            joint_torques=qfrc_inverse,
-            success=True,
-            is_feasible=True,
-            # Fill validation metrics if available
-        )
-
-    def compute_torques_with_posture(
-        self,
-        qpos: np.ndarray,
-        qvel: np.ndarray,
-        qacc_primary: np.ndarray,
-        qpos_desired: np.ndarray,
-        kp_posture: float = 10.0,
-        primary_body_name: str = "club_head",
-    ) -> InverseDynamicsResult:
-        """Compute torques achieving primary task + secondary posture (Phase 4).
-
-        Uses Null-Space Projection:
-            tau_total = tau_primary + (I - J^T(J J^T)^-1 J) * tau_secondary
-
-        This ensures secondary tasks (like posture) do not interfere with the
-        primary task (e.g. club head trajectory).
-
-        Args:
-            qpos: Current joint positions
-            qvel: Current joint velocities
-            qacc_primary: Desired accelerations for the primary task
-            qpos_desired: Target posture configuration
-            kp_posture: Gain for posture control
-            primary_body_name: Name of the body representing the primary task
-
-        Returns:
-            InverseDynamicsResult with combined torques
-        """
-        # 1. Compute Primary Task Torques (using standard Inverse Dynamics)
-        # Note: This assumes qacc_primary satisfies the task constraints
-        if not (qpos is not None):
-            raise ValueError("qpos must be provided")
-        if not (qpos is not None):
-            raise ValueError("qpos must be provided")
-        primary_result = self.compute_required_torques(qpos, qvel, qacc_primary)
-        tau_primary = primary_result.joint_torques  # Total generalized force
-
-        # 2. Compute Jacobian for Primary Task
-        body_id = mujoco.mj_name2id(
-            self.model, mujoco.mjtObj.mjOBJ_BODY, primary_body_name
-        )
-        if body_id == -1:
-            # Fallback or error? For now log warning and treat as no task
-            logger.warning(
-                f"Primary body '{primary_body_name}' not found. Using pure posture."
-            )
-            J_primary = np.zeros((3, self.model.nv))
-        else:
-            jacp = np.zeros((3, self.model.nv))
-            jacr = np.zeros((3, self.model.nv))
-            # Use private data state which matches qpos
-            mujoco.mj_jacBody(self.model, self._perturb_data, jacp, jacr, body_id)
-            # Combine pos + rot jacobian? Usually just pos for hitting ball
-            J_primary = jacp
-
-        # 3. Compute Null-Space Projector: N = I - J^+ J
-        # Pinverse: J^+ = J^T (J J^T)^-1  (for full rank)
-        # We use numpy's pinv for safety
-        J_pinv = np.linalg.pinv(J_primary)
-        N = np.eye(self.model.nv) - np.dot(J_pinv, J_primary)
-
-        # 4. Compute Secondary Posture Torques (PD control in joint space)
-        # tau_posture = Kp * (q_des - q) - Kd * q_vel
-        q_err = qpos_desired - qpos
-        # Simple PD
-        kd_posture = 2 * np.sqrt(kp_posture)  # Critical damping approx
-        tau_secondary = (kp_posture * q_err) - (kd_posture * qvel)
-
-        # 5. Project Secondary into Null Space
-        tau_null = np.dot(N, tau_secondary)
-
-        # 6. Combine
-        tau_total = tau_primary + tau_null
-
-        return InverseDynamicsResult(
-            joint_torques=tau_total,
-            success=True,
-            is_feasible=True,
-            manipulability_index=primary_result.manipulability_index,
-        )
-
-    def _compute_gravity_force(self) -> np.ndarray:
-        qvel_backup = self._perturb_data.qvel.copy()
-        self._perturb_data.qvel[:] = 0
-        self._perturb_data.cvel[:] = 0
-        g_force = np.zeros(self.model.nv)
-        mujoco.mj_rne(self.model, self._perturb_data, 0, g_force)
-        self._perturb_data.qvel[:] = qvel_backup
-        return g_force
-
-    def _compute_coriolis_force(self, g_force: np.ndarray) -> np.ndarray:
-        if not (g_force is not None):
-            raise ValueError("g_force must be provided")
-        if not (g_force is not None):
-            raise ValueError("g_force must be provided")
-        mujoco.mj_forward(self.model, self._perturb_data)
-        bias_force = self._perturb_data.qfrc_bias.copy()
-        return bias_force - g_force
-
-    def _compute_control_force(self, ctrl: np.ndarray) -> np.ndarray:
-        if not (ctrl is not None):
-            raise ValueError("ctrl must be provided")
-        if not (ctrl is not None):
-            raise ValueError("ctrl must be provided")
-        self._perturb_data.ctrl[:] = 0
-        if len(ctrl) == self.model.nu:
-            self._perturb_data.ctrl[:] = ctrl
-        mujoco.mj_fwdActuation(self.model, self._perturb_data)
-        return self._perturb_data.qfrc_actuation.copy()
-
-    def _solve_component_accelerations(
-        self,
-        g_force: np.ndarray,
-        c_force: np.ndarray,
-        tau_force: np.ndarray,
-    ) -> InducedAccelerationResult:
-        # Acc_G = M^-1 * (-G), Acc_C = M^-1 * (-C), Acc_Tau = M^-1 * (tau)
-        if not (g_force is not None):
-            raise ValueError("g_force must be provided")
-        if not (g_force is not None):
-            raise ValueError("g_force must be provided")
-        a_g = (-g_force).copy()
-        mujoco.mj_solveM(self.model, self._perturb_data, a_g)
-
-        a_c = (-c_force).copy()
-        mujoco.mj_solveM(self.model, self._perturb_data, a_c)
-
-        a_t = tau_force.copy()
-        mujoco.mj_solveM(self.model, self._perturb_data, a_t)
-
-        total = a_g + a_c + a_t
-        return InducedAccelerationResult(
-            gravity=a_g, velocity=a_c, control=a_t, total=total
-        )
-
-    @precondition(
-        lambda self, qpos, qvel, ctrl: len(qpos) > 0,
-        "Joint positions must be non-empty",
-    )
-    def compute_induced_accelerations(
-        self,
-        qpos: np.ndarray,
-        qvel: np.ndarray,
-        ctrl: np.ndarray,
-    ) -> InducedAccelerationResult:
-        """Compute acceleration components induced by different forces.
-
-        Using M(q)q_ddot = tau - C(q,q_dot)q_dot - G(q)
-        q_ddot = M^-1 * (tau - C - G)
-
-        FIXED (Assessment A Finding A-001 - CRITICAL):
-        Now uses self._perturb_data for thread safety instead of modifying
-        shared self.data state. This prevents race conditions during parallel
-        analysis where visualization thread reads from self.data.
-
-        Args:
-            qpos: Joint positions [nv]
-            qvel: Joint velocities [nv]
-            ctrl: Applied control torques [nu] (or nv if full actuation)
-
-        Returns:
-            InducedAccelerationResult with component accelerations.
-        """
-        if not (qpos is not None):
-            raise ValueError("qpos must be provided")
-        if not (qpos is not None):
-            raise ValueError("qpos must be provided")
-        self._perturb_data.qpos[:] = qpos
-        self._perturb_data.qvel[:] = qvel
-
-        g_force = self._compute_gravity_force()
-        c_force = self._compute_coriolis_force(g_force)
-        tau_force = self._compute_control_force(ctrl)
-
-        return self._solve_component_accelerations(g_force, c_force, tau_force)
-
-    @precondition(
-        lambda self, times, positions, velocities, accelerations: len(times) > 0,
-        "Time array must be non-empty",
-    )
-    @precondition(
-        lambda self, times, positions, velocities, accelerations: (
-            len(times) == len(positions) == len(velocities) == len(accelerations)
-        ),
-        "All trajectory arrays must have the same length",
-    )
-    def solve_inverse_dynamics_trajectory(
-        self,
-        times: np.ndarray,
-        positions: np.ndarray,
-        velocities: np.ndarray,
-        accelerations: np.ndarray,
-    ) -> list[InverseDynamicsResult]:
-        """Solve inverse dynamics for entire trajectory.
-
-        Args:
-            times: Time array [N]
-            positions: Joint positions [N x nv]
-            velocities: Joint velocities [N x nv]
-            accelerations: Joint accelerations [N x nv]
-
-        Returns:
-            List of InverseDynamicsResult for each time step
-        """
-        if not (times is not None):
-            raise ValueError("times must be provided")
-        if not (times is not None):
-            raise ValueError("times must be provided")
-        results = []
-
-        for i in range(len(times)):
-            result = self.compute_required_torques(
-                positions[i],
-                velocities[i],
-                accelerations[i],
-            )
-            results.append(result)
-
-        return results
-
-    def compute_partial_inverse_dynamics(
-        self,
-        qpos: np.ndarray,
-        qvel: np.ndarray,
-        qacc: np.ndarray,
-        constrained_joints: list[int],
-    ) -> InverseDynamicsResult:
-        """Compute partial inverse dynamics for parallel mechanisms.
-
-        For closed-chain systems, some joints may be constrained.
-        This computes torques for actuated joints while respecting
-        constraint forces.
-
-        Args:
-            qpos: Joint positions [nv]
-            qvel: Joint velocities [nv]
-            qacc: Joint accelerations [nv]
-            constrained_joints: List of constrained joint indices
-
-        Returns:
-            InverseDynamicsResult with partial solution
-        """
-        # Full inverse dynamics
-        if not (qpos is not None):
-            raise ValueError("qpos must be provided")
-        if not (qpos is not None):
-            raise ValueError("qpos must be provided")
-        full_result = self.compute_required_torques(qpos, qvel, qacc)
-
-        # Create selection matrix for actuated joints
-        actuated_joints = [
-            i for i in range(self.model.nv) if i not in constrained_joints
-        ]
-
-        # Extract actuated torques
-        full_result.joint_torques[actuated_joints]
-
-        # For constrained joints, torques come from constraints
-        np.zeros(len(constrained_joints))
-        if full_result.constraint_forces is not None:
-            full_result.constraint_forces[constrained_joints]
-
-        return full_result  # Return full result with constraint info
-
-    def decompose_forces(
-        self,
-        qpos: np.ndarray,
-        qvel: np.ndarray,
-        qacc: np.ndarray,
-    ) -> ForceDecomposition:
-        """Decompose total forces into components.
-
-        Args:
-            qpos: Joint positions [nv]
-            qvel: Joint velocities [nv]
-            qacc: Joint accelerations [nv]
-
-        Returns:
-            ForceDecomposition with all components
-        """
-        if not (qpos is not None):
-            raise ValueError("qpos must be provided")
-        if not (qpos is not None):
-            raise ValueError("qpos must be provided")
-        result = self.compute_required_torques(qpos, qvel, qacc)
-
-        # Decompose Coriolis into centrifugal
-        centrifugal, _ = self.kinematic_analyzer.decompose_coriolis_forces(qpos, qvel)
-
-        # Provide defaults for None values
-        nv = len(result.joint_torques)
-        inertial = (
-            result.inertial_torques
-            if result.inertial_torques is not None
-            else np.zeros(nv)
-        )
-        coriolis = (
-            result.coriolis_torques
-            if result.coriolis_torques is not None
-            else np.zeros(nv)
-        )
-        gravity = (
-            result.gravity_torques
-            if result.gravity_torques is not None
-            else np.zeros(nv)
-        )
-
-        return ForceDecomposition(
-            total=result.joint_torques,
-            inertial=inertial,
-            coriolis=coriolis,
-            centrifugal=centrifugal,
-            gravity=gravity,
-        )
-
-    @precondition(
-        lambda self, qpos, qvel, qacc, body_id: body_id >= 0,
-        "Body ID must be non-negative",
-    )
-    def compute_end_effector_forces(
-        self,
-        qpos: np.ndarray,
-        qvel: np.ndarray,
-        qacc: np.ndarray,
-        body_id: int,
-    ) -> np.ndarray:
-        """Compute forces at end-effector (e.g., club head).
-
-        Maps joint torques to task-space forces: F = (J^T)^{-1} τ
-
-        PERFORMANCE NOTE (Issue B-003): Uses lstsq which is correct but not optimized.
-        For batch processing of trajectories, consider precomputing pseudo-inverses
-        if the robot configuration remains similar. However, since Jacobian depends
-        on qpos (configuration-dependent), caching is only beneficial for repeated
-        calls with identical qpos but different torques (rare in practice).
-
-        Potential optimization for batch processing:
-            # For trajectory analysis (same qpos, varying torques):
-            J_pinv = np.linalg.pinv(jacp.T)  # Compute once
-            ee_forces = J_pinv @ torques_batch  # Reuse for multiple torques
-
-        Args:
-            qpos: Joint positions [nv]
-            qvel: Joint velocities [nv]
-            qacc: Joint accelerations [nv]
-            body_id: Body ID for end-effector
-
-        Returns:
-            End-effector force [3]
-        """
-        # Compute required torques
-        if not (qpos is not None):
-            raise ValueError("qpos must be provided")
-        if not (qpos is not None):
-            raise ValueError("qpos must be provided")
-        result = self.compute_required_torques(qpos, qvel, qacc)
-
-        # Get Jacobian (configuration-dependent, must recompute for each qpos)
-        self.data.qpos[:] = qpos
-        mujoco.mj_forward(self.model, self.data)
-
-        # Compute Jacobian using pre-allocated arrays and detected API
-        if self._use_flat_jacobian:
-            mujoco.mj_jacBody(
-                self.model,
-                self.data,
-                self._jacp_flat,
-                self._jacr_flat,
-                body_id,
-            )
-            jacp = self._jacp_flat.reshape(3, self.model.nv)
-        else:
-            mujoco.mj_jacBody(self.model, self.data, self._jacp, self._jacr, body_id)
-            jacp = self._jacp
-
-        # Map torques to forces: F = (J^T)^{-1} τ
-        # lstsq is robust for redundant/constrained systems (handles rank deficiency)
-        ee_force, _residuals, _rank, _s = lstsq(jacp.T, result.joint_torques)
-
-        return np.array(ee_force, dtype=np.float64)
-
-    def validate_solution(
-        self,
-        qpos: np.ndarray,
-        qvel: np.ndarray,
-        qacc: np.ndarray,
-        computed_torques: np.ndarray,
-    ) -> dict[str, float]:
-        """Validate inverse dynamics solution.
-
-        Checks if computed torques actually produce desired acceleration.
-
-        FIXED: This method now uses MjDataContext for state isolation and
-        static calculation (mj_forward) instead of mj_step to avoid the
-        "Observer Effect" bug where validation would advance simulation time
-        and corrupt subsequent calculations.
-        See Issues A-003 and F-001.
-
-        Args:
-            qpos: Joint positions [nv]
-            qvel: Joint velocities [nv]
-            qacc: Desired accelerations [nv]
-            computed_torques: Computed torques [nv]
-
-        Returns:
-            Validation metrics
-        """
-        # Use context manager for automatic state save/restore (Issues A-003, F-001)
-        with MjDataContext(self.model, self.data):
-            # Apply torques in forward dynamics
-            self.data.qpos[:] = qpos
-            self.data.qvel[:] = qvel
-            self.data.ctrl[: self.model.nu] = computed_torques[: self.model.nu]
-
-            # Compute forward dynamics (static calculation, no time advancement)
-            # FIXED: Removed mj_step which was causing Observer Effect
-            mujoco.mj_forward(self.model, self.data)
-
-            # Get resulting acceleration
-            m_matrix = np.zeros((self.model.nv, self.model.nv))
-            mujoco.mj_fullM(self.model, m_matrix, self.data.qM)
-
-            # Acceleration from dynamics: M^{-1}(τ - C q̇ - g)
-            coriolis = self.kinematic_analyzer.compute_coriolis_forces(qpos, qvel)
-            gravity = self.kinematic_analyzer.compute_gravity_forces(qpos)
-
-            m_inv = np.linalg.inv(m_matrix)
-            computed_qacc = m_inv @ (computed_torques - coriolis - gravity)
-
-            # Error metrics
-            acc_error = np.linalg.norm(computed_qacc - qacc)
-            relative_error = acc_error / (np.linalg.norm(qacc) + 1e-10)
-
-            return {
-                "acceleration_error": float(acc_error),
-                "relative_error": float(relative_error),
-                "max_torque": float(np.max(np.abs(computed_torques))),
-                "mean_torque": float(np.mean(np.abs(computed_torques))),
-            }
-        # State is automatically restored here by context manager
-
-    def compute_actuator_efficiency(
-        self,
-        result: InverseDynamicsResult,
-    ) -> dict[str, float]:
-        """Compute efficiency metrics for actuators.
-
-        Args:
-            result: Inverse dynamics result
-
-        Returns:
-            Efficiency metrics
-        """
-        if not (result is not None):
-            raise ValueError("result must be provided")
-        if not (result is not None):
-            raise ValueError("result must be provided")
-        torques = result.joint_torques
-
-        # Mechanical advantage (ratio of output to input)
-        if result.inertial_torques is not None:
-            inertial_ratio = float(
-                np.linalg.norm(result.inertial_torques)
-                / (np.linalg.norm(torques) + 1e-10),
-            )
-        else:
-            inertial_ratio = 0.0
-
-        # Gravity compensation ratio
-        if result.gravity_torques is not None:
-            gravity_ratio = float(
-                np.linalg.norm(result.gravity_torques)
-                / (np.linalg.norm(torques) + 1e-10),
-            )
-        else:
-            gravity_ratio = 0.0
-
-        # Coriolis ratio (ideally small)
-        if result.coriolis_torques is not None:
-            coriolis_ratio = float(
-                np.linalg.norm(result.coriolis_torques)
-                / (np.linalg.norm(torques) + 1e-10),
-            )
-        else:
-            coriolis_ratio = 0.0
-
-        return {
-            "inertial_ratio": float(inertial_ratio),
-            "gravity_ratio": float(gravity_ratio),
-            "coriolis_ratio": float(coriolis_ratio),
-            "efficiency_index": float(
-                inertial_ratio / (gravity_ratio + coriolis_ratio + 1e-10),
-            ),
-        }
-
-
-class RecursiveNewtonEuler:
-    """Recursive Newton-Euler algorithm for inverse dynamics.
-
-    More efficient than matrix-based approach for serial chains.
-    Useful for real-time applications.
-    """
-
-    def __init__(self, model: mujoco.MjModel, data: mujoco.MjData) -> None:
-        """Initialize RNE solver.
-
-        Args:
-            model: MuJoCo model
-            data: MuJoCo data
-        """
-        if not (model is not None):
-            raise ValueError("model must be provided")
-        if not (model is not None):
-            raise ValueError("model must be provided")
-        self.model = model
-        self.data = data
-
-    def compute(
-        self,
-        qpos: np.ndarray,
-        qvel: np.ndarray,
-        qacc: np.ndarray,
-    ) -> np.ndarray:
-        """Compute inverse dynamics using RNE.
-
-        Args:
-            qpos: Joint positions [nv]
-            qvel: Joint velocities [nv]
-            qacc: Joint accelerations [nv]
-
-        Returns:
-            Joint torques [nv]
-        """
-        # MuJoCo's internal RNE is very efficient
-        # We use MuJoCo's inverse dynamics
-        if not (qpos is not None):
-            raise ValueError("qpos must be provided")
-        if not (qpos is not None):
-            raise ValueError("qpos must be provided")
-        self.data.qpos[:] = qpos
-        self.data.qvel[:] = qvel
-        self.data.qacc[:] = qacc
-
-        # Use MuJoCo's rne function
-        result = np.zeros(self.model.nv)
-        mujoco.mj_rne(self.model, self.data, 0, result)
-
-        return result
-
-
-def _validate_inverse_dynamics_export_inputs(
-    times: np.ndarray,
-    results: list[InverseDynamicsResult],
-) -> int:
-    """Validate inputs for inverse dynamics CSV export.
-
-    Args:
-        times: Time array [N].
-        results: List of InverseDynamicsResult.
-
-    Returns:
-        Number of joints (nv) across all results.
-
-    Raises:
-        TypeError: If inputs are wrong type.
-        ValueError: If inputs are empty or inconsistent.
-    """
-    if not isinstance(times, np.ndarray):
-        raise TypeError(f"times must be numpy array, got {type(times).__name__}")
-
-    if not isinstance(results, list):
-        raise TypeError(f"results must be list, got {type(results).__name__}")
-
-    if len(results) == 0:
-        raise ValueError("Cannot export empty results list")
-
-    if len(times) != len(results):
-        raise ValueError(
-            f"Length mismatch: times has {len(times)} elements, "
-            f"results has {len(results)} elements"
-        )
-
-    for i, result in enumerate(results):
-        if not isinstance(result, InverseDynamicsResult):
-            raise TypeError(
-                f"results[{i}] is {type(result).__name__}, "
-                f"expected InverseDynamicsResult"
-            )
-
-    nv = len(results[0].joint_torques)
-    for i, result in enumerate(results):
-        if len(result.joint_torques) != nv:
-            raise ValueError(
-                f"Inconsistent joint count: results[0] has {nv} joints, "
-                f"results[{i}] has {len(result.joint_torques)} joints"
-            )
-    return nv
-
-
-def _build_inverse_dynamics_csv_row(
-    result: InverseDynamicsResult, time_val: float, nv: int
-) -> list[float]:
-    """Build a single CSV row from an inverse dynamics result.
-
-    Args:
-        result: Single timestep result.
-        time_val: Time value for this row.
-        nv: Number of joints.
-
-    Returns:
-        List of float values for the CSV row.
-    """
-    if not (result is not None):
-        raise ValueError("result must be provided")
-    if not (result is not None):
-        raise ValueError("result must be provided")
-    row: list[float] = [time_val]
-    for i in range(nv):
-        row.append(result.joint_torques[i])
-        row.append(
-            result.inertial_torques[i] if result.inertial_torques is not None else 0.0
-        )
-        row.append(
-            result.coriolis_torques[i] if result.coriolis_torques is not None else 0.0
-        )
-        row.append(
-            result.gravity_torques[i] if result.gravity_torques is not None else 0.0
-        )
-    row.append(result.residual_norm)
-    return row
-
-
-def export_inverse_dynamics_to_csv(
-    times: np.ndarray,
-    results: list[InverseDynamicsResult],
-    filepath: str,
-) -> None:
-    """Export inverse dynamics results to CSV.
-
-    Args:
-        times: Time array [N]
-        results: List of InverseDynamicsResult
-        filepath: Output CSV path
-
-    Raises:
-        ValueError: If times and results have mismatched lengths
-        TypeError: If results contains non-InverseDynamicsResult items
-        ValueError: If results list is empty
-
-    Note:
-        FIXED per Assessment A Finding A-007: Added comprehensive input
-        validation to prevent malformed CSV output and silent failures.
-    """
-    if not (times is not None):
-        raise ValueError("times must be provided")
-    if not (times is not None):
-        raise ValueError("times must be provided")
-    nv = _validate_inverse_dynamics_export_inputs(times, results)
-
-    with open(filepath, "w", newline="") as f:
-        writer = csv.writer(f)
-
-        # Header
-        header = ["time"]
-        for i in range(nv):
-            header.extend(
-                [f"torque_{i}", f"inertial_{i}", f"coriolis_{i}", f"gravity_{i}"],
-            )
-        header.append("residual_norm")
-        writer.writerow(header)
-
-        # Data rows
-        for t, result in zip(times, results, strict=False):
-            writer.writerow(_build_inverse_dynamics_csv_row(result, t, nv))
-
-
-class InverseDynamicsAnalyzer:
-    """High-level analyzer combining inverse dynamics and kinematic forces.
-
-    This class provides the complete analysis pipeline for understanding
-    swing dynamics from motion capture data.
-    """
-
-    def __init__(self, model: mujoco.MjModel, data: mujoco.MjData) -> None:
-        """Initialize analyzer.
-
-        Args:
-            model: MuJoCo model
-            data: MuJoCo data
-        """
-        if not (model is not None):
-            raise ValueError("model must be provided")
-        if not (model is not None):
-            raise ValueError("model must be provided")
-        self.id_solver = InverseDynamicsSolver(model, data)
-        self.kin_analyzer = KinematicForceAnalyzer(model, data)
-
-    def analyze_captured_motion(
-        self,
-        times: np.ndarray,
-        positions: np.ndarray,
-        velocities: np.ndarray,
-        accelerations: np.ndarray,
-    ) -> dict:
-        """Complete analysis of captured motion.
-
-        This is the main method for analyzing motion capture data.
-        Computes both kinematic forces (Coriolis, centrifugal) and
-        inverse dynamics (required torques).
-
-        Args:
-            times: Time array [N]
-            positions: Joint positions [N x nv]
-            velocities: Joint velocities [N x nv]
-            accelerations: Joint accelerations [N x nv]
-
-        Returns:
-            Dictionary with comprehensive analysis
-        """
-        # Kinematic force analysis
-        if not (times is not None):
-            raise ValueError("times must be provided")
-        if not (times is not None):
-            raise ValueError("times must be provided")
-        kinematic_forces = self.kin_analyzer.analyze_trajectory(
-            times,
-            positions,
-            velocities,
-            accelerations,
-        )
-
-        # Inverse dynamics
-        id_results = self.id_solver.solve_inverse_dynamics_trajectory(
-            times,
-            positions,
-            velocities,
-            accelerations,
-        )
-
-        # Aggregate statistics
-        peak_coriolis_power = 0.0
-        max_joint_torque = 0.0
-
-        for kf in kinematic_forces:
-            peak_coriolis_power = max(peak_coriolis_power, abs(kf.coriolis_power))
-
-        for id_res in id_results:
-            max_joint_torque = max(
-                max_joint_torque,
-                np.max(np.abs(id_res.joint_torques)),
-            )
-
-        return {
-            "kinematic_forces": kinematic_forces,
-            "inverse_dynamics": id_results,
-            "statistics": {
-                "peak_coriolis_power": peak_coriolis_power,
-                "max_joint_torque": max_joint_torque,
-                "duration": times[-1] - times[0],
-                "num_frames": len(times),
-            },
-        }
-
-    def compare_swings(self, swing1_data: dict, swing2_data: dict) -> dict:
-        """Compare two swing analyses.
-
-        Args:
-            swing1_data: First swing analysis
-            swing2_data: Second swing analysis
-
-        Returns:
-            Comparison metrics
-        """
-        if not (swing1_data is not None):
-            raise ValueError("swing1_data must be provided")
-        if not (swing1_data is not None):
-            raise ValueError("swing1_data must be provided")
-        stats1 = swing1_data["statistics"]
-        stats2 = swing2_data["statistics"]
-
-        return {
-            "coriolis_power_diff": stats2["peak_coriolis_power"]
-            - stats1["peak_coriolis_power"],
-            "torque_diff": stats2["max_joint_torque"] - stats1["max_joint_torque"],
-            "duration_diff": stats2["duration"] - stats1["duration"],
-        }
+__all__ = [
+    "ForceDecomposition",
+    "InducedAccelerationResult",
+    "InverseDynamicsAnalyzer",
+    "InverseDynamicsResult",
+    "InverseDynamicsSolver",
+    "RecursiveNewtonEuler",
+    "_build_inverse_dynamics_csv_row",
+    "_validate_inverse_dynamics_export_inputs",
+    "export_inverse_dynamics_to_csv",
+]
